update directory entity, daos e dtos
cambiata directory entity, daos e dtos
