+++ src/main/java/com/develhope/spring/models/dtos/ProductTypeDto.java
@@ -1,0 +1,32 @@
+package com.develhope.spring.models.dtos;
+
+public class ProductTypeDto {
+
+    private String id;
+
+    private String productType;
+
+    public ProductTypeDto() {
+    }
+
+    public ProductTypeDto(String id, String productType) {
+        this.id = id;
+        this.productType = productType;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getProductType() {
+        return productType;
+    }
+
+    public void setProductType(String productType) {
+        this.productType = productType;
+    }
+}
