+++ src/main/java/com/develhope/spring/models/entities/ProductEntity.java
@@ -1,0 +1,80 @@
+package com.develhope.spring.models.entities;
+
+import jakarta.persistence.*;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+@Entity
+@Table(name = "products")
+public class ProductEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private String id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private BigDecimal price;
+
+    private String ingredients;
+
+    @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinTable(name = "product_productType", joinColumns = @JoinColumn(name = "product_id"),
+            inverseJoinColumns = @JoinColumn(name = "productType_id"))
+    @Column(name = "productTypes")
+    private List<ProductTypeEntity> productTypes;
+
+    public ProductEntity() {
+    }
+
+    public ProductEntity(String id, String name, BigDecimal price, String ingredients, List<ProductTypeEntity> productTypes) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.ingredients = ingredients;
+        this.productTypes = productTypes;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
+
+    public String getIngredients() {
+        return ingredients;
+    }
+
+    public void setIngredients(String ingredients) {
+        this.ingredients = ingredients;
+    }
+
+    public List<ProductTypeEntity> getProductTypes() {
+        return productTypes;
+    }
+
+    public void setProductTypes(List<ProductTypeEntity> productTypes) {
+        this.productTypes = productTypes;
+    }
+}
