--- src/main/java/com/develhope/spring/repositories/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.develhope.spring.repositories;
-
-import com.develhope.spring.entities.ProductEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, UUID> {
-}
