--- src/main/java/com/develhope/spring/repositories/ProductTypeRepository.java
@@ -1,11 +1,0 @@
-package com.develhope.spring.repositories;
-
-import com.develhope.spring.entities.ProductTypeEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface ProductTypeRepository extends JpaRepository<ProductTypeEntity, UUID> {
-}
